fcm: rename command 'cluster' to 'cluster create'
